app/requirements.txt
flask
celery[redis]
flask-restplus
flask-mongoengine
pipenv
flask-testing
flask-script
marshmallow
uwsgi